[feature/12] Add lessons query (#45)
* Add Domain project with Themes and Lessons

* Implement dynamic argument extraction

* Add mock data to lesson repository

* Fix rebase

* Rename argument

* Drop generic repository interface

* Simplify repository filter methods

* Fix type of theme id in filter

* Delete unused imports

* Convert lesson statuses factory to repository

* Move mock data for step out of query

* Dynamically load repositories

* Cleanup context extensions

* Add repository implementation in the Data layer

* Extend DB model

* Configure how duration is converted from the database

* Use entities that derive from models

Not convinced this is a good approach.
Alternative is to more verbose map from
entities to the models, but have to consider how we load associations.
Todo this later.

* Use real repositories

Side effect is to make repositories scoped to match the db context scope. Ideally we would have them still singleton and figure out a way to inject the scoped context since nothing would run outside a scope

* Add how to run

* Update context and add converter for id

* Make ids strings and generate them client side

* Seed dummy data

* Include thumnail for theme

* Drop special method for cursor